+++ src/ops/bootstrap.js
@@ -86,7 +86,6 @@ import * as messagePort from "ext:deno_web/13_message_port.js";
 import * as compression from "ext:deno_web/14_compression.js";
 import * as performance from "ext:deno_web/15_performance.js";
 import * as imageData from "ext:deno_web/16_image_data.js";
-import * as caches from "ext:deno_cache/01_cache.js";
 
 // deno_web (abortSignal)
 propNonEnumerable("AbortController", abortSignal.AbortController);
@@ -416,6 +415,19 @@ propNonEnumerableLazyLoaded(
 );
 
 /** deno_cache */
+import * as caches from "ext:deno_cache/01_cache.js";
+
+propGetterOnly("caches", caches.cacheStorage);
+propNonEnumerable("CacheStorage", caches.CacheStorage);
+propNonEnumerable("Cache", caches.Cache);
+
+/** deno_websocket */
+import * as webSocket from "ext:deno_websocket/01_websocket.js";
+import * as webSocketStream from "ext:deno_websocket/02_websocketstream.js";
+
+propNonEnumerable("WebSocket", webSocket.WebSocket);
+propNonEnumerable("WebSocketStream", webSocketStream.WebSocketStream);
+propNonEnumerable("WebSocketError", webSocketStream.WebSocketError);
 
 /** toxo_navigator */
 import * as nav from "ext:toxo_navigator/navigator.js";
@@ -432,10 +444,6 @@ propWritable("prompt", prompt.prompt);
 
 /** toxo_filesystem */
 import * as fs from "ext:toxo_filesystem/filesystem.js";
-
-propGetterOnly("caches", caches.cacheStorage);
-propNonEnumerable("CacheStorage", caches.CacheStorage);
-propNonEnumerable("Cache", caches.Cache);
 
 propNonEnumerable("StorageManager", fs.StorageManager);
 propNonEnumerable("FileSystemHandle", fs.FileSystemHandle);